Add abstracted sign in popup widget
